+++ components/atoms/Card/card.tsx
@@ -4,16 +4,18 @@ interface CardProps {
   className?: string;
   heading?: ReactNode;
   children: ReactNode;
+  focusable?: boolean;
 }
 
-const Card: React.FC<CardProps> = ({ className, children, heading }) => {
+const Card = ({ className, children, heading, focusable = false }: CardProps) => {
   return (
     <article
       className={`${className ? className : ""} block ${heading ? "" : "p-3"} bg-white border rounded-lg shadow-xs`}
+      tabIndex={focusable ? 0 : undefined}
     >
       {heading ? (
         <>
-          <div className="px-3 md:px-6 py-3 rounded-t-lg bg-light-slate-3">{heading}</div>
+          <div className="px-3 py-3 rounded-t-lg md:px-6 bg-light-slate-3">{heading}</div>
           <div>{children}</div>
         </>
       ) : (