+++ layouts/lists.tsx
@@ -20,6 +20,7 @@ const ListPageLayout = ({
   repoFilter = false,
   repoFilterOptions = [],
   repoFilterSelect = () => {},
+  overLimit,
 }: {
   children: React.ReactNode;
   list?: DBList;
@@ -31,6 +32,7 @@ const ListPageLayout = ({
   repoFilter?: boolean;
   repoFilterOptions?: OptionKeys[];
   repoFilterSelect?: (repo: OptionKeys[]) => void;
+  overLimit?: boolean;
 }) => {
   const router = useRouter();
   const { range } = router.query;
@@ -56,6 +58,7 @@ const ListPageLayout = ({
               workspaceId={workspaceId}
               isOwner={isOwner}
               owners={owners}
+              overLimit={overLimit}
             />
           )}
         </Header>