+++ pages/workspaces/[workspaceId]/repository-insights/[insightId]/edit.tsx
@@ -13,9 +13,11 @@ import { WorkspaceLayout } from "components/Workspaces/WorkspaceLayout";
 import { TrackedReposTable } from "components/Workspaces/TrackedReposTable";
 import Title from "components/atoms/Typography/title";
 import Text from "components/atoms/Typography/text";
+import { useIsWorkspaceUpgraded } from "lib/hooks/api/useIsWorkspaceUpgraded";
 
 const TrackedReposModal = dynamic(import("components/Workspaces/TrackedReposModal"));
 const DeleteInsightPageModal = dynamic(import("components/organisms/InsightPage/DeleteInsightPageModal"));
+const InsightUpgradeModal = dynamic(() => import("components/Workspaces/InsightUpgradeModal"));
 
 export async function getServerSideProps(context: GetServerSidePropsContext) {
   const supabase = createPagesServerClient(context);
@@ -91,6 +93,9 @@ export default function RepoInsightEditPage({ insight, workspaceId, isOwner, bea
   const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
   const [trackedReposModalOpen, setTrackedReposModalOpen] = useState(false);
   const [trackedRepos, setTrackedRepos] = useState<Map<string, boolean>>(initialTrackedRepos);
+  const [isInsightUpgradeModalOpen, setIsInsightUpgradeModalOpen] = useState(false);
+  const { data: isWorkspaceUpgraded } = useIsWorkspaceUpgraded({ workspaceId });
+  const showBanner = isOwner && !isWorkspaceUpgraded;
 
   const updateInsight = async () => {
     setLoading(true);
@@ -171,6 +176,10 @@ export default function RepoInsightEditPage({ insight, workspaceId, isOwner, bea
           disabled={loading}
           repositories={trackedRepos}
           onAddRepos={() => {
+            if (showBanner) {
+              setIsInsightUpgradeModalOpen(true);
+              return;
+            }
             setTrackedReposModalOpen(true);
           }}
           onRemoveTrackedRepo={(event) => {
@@ -239,6 +248,15 @@ export default function RepoInsightEditPage({ insight, workspaceId, isOwner, bea
           setOpen={() => setIsDeleteModalOpen(true)}
           onClose={() => setIsDeleteModalOpen(false)}
           onConfirm={deleteInsight}
+        />
+      ) : null}
+      {isOwner ? (
+        <InsightUpgradeModal
+          workspaceId={workspaceId}
+          variant="all"
+          isOpen={isInsightUpgradeModalOpen}
+          onClose={() => setIsInsightUpgradeModalOpen(false)}
+          overLimit={trackedRepos.size}
         />
       ) : null}
     </WorkspaceLayout>